move each runtime to its test folder
