+++ tests/integration/buildcmd/build_integ_base.py
@@ -1,34 +1,26 @@
+import json
 import os
 import posixpath
-import sys
-import uuid
 import shutil
 import tempfile
 import time
-import logging
-import json
+import uuid
+from pathlib import Path
 from typing import Optional
 from unittest import TestCase
 
 import docker
 import jmespath
-from pathlib import Path
 
-from samcli.commands.build.utils import MountMode
-from samcli.lib.utils import osutils
-from samcli.lib.utils.architecture import X86_64, has_runtime_multi_arch_image
+from samcli.lib.utils.architecture import X86_64
 from samcli.local.docker.lambda_build_container import LambdaBuildContainer
 from samcli.yamlhelper import yaml_parse
 from tests.testing_utils import (
     IS_WINDOWS,
     run_command,
     SKIP_DOCKER_TESTS,
-    SKIP_DOCKER_MESSAGE,
-    SKIP_DOCKER_BUILD,
     get_sam_command,
 )
-
-LOG = logging.getLogger(__name__)
 
 
 class BuildIntegBase(TestCase):
@@ -239,578 +231,6 @@ class BuildIntegBase(TestCase):
         if architecture:
             overrides["Architectures"] = architecture
         return overrides
-
-
-class BuildIntegRubyBase(BuildIntegBase):
-    EXPECTED_FILES_PROJECT_MANIFEST = {"app.rb"}
-    EXPECTED_RUBY_GEM = "aws-record"
-
-    FUNCTION_LOGICAL_ID = "Function"
-
-    def _test_with_default_gemfile(self, runtime, use_container, code_uri, relative_path, architecture=None):
-        overrides = self.get_override(runtime, code_uri, architecture, "ignored")
-        cmdlist = self.get_command_list(use_container=use_container, parameter_overrides=overrides)
-
-        LOG.info("Running Command:")
-        LOG.info(cmdlist)
-        run_command(cmdlist, cwd=self.working_dir)
-
-        self._verify_built_artifact(
-            self.default_build_dir,
-            self.FUNCTION_LOGICAL_ID,
-            self.EXPECTED_FILES_PROJECT_MANIFEST,
-            self.EXPECTED_RUBY_GEM,
-        )
-
-        self._verify_resource_property(
-            str(self.built_template),
-            "OtherRelativePathResource",
-            "BodyS3Location",
-            os.path.relpath(
-                os.path.normpath(os.path.join(str(relative_path), "SomeRelativePath")),
-                str(self.default_build_dir),
-            ),
-        )
-
-        self._verify_resource_property(
-            str(self.built_template),
-            "GlueResource",
-            "Command.ScriptLocation",
-            os.path.relpath(
-                os.path.normpath(os.path.join(str(relative_path), "SomeRelativePath")),
-                str(self.default_build_dir),
-            ),
-        )
-
-        if use_container:
-            self.verify_docker_container_cleanedup(runtime)
-            self.verify_pulled_image(runtime, architecture)
-
-    def _verify_built_artifact(self, build_dir, function_logical_id, expected_files, expected_modules):
-        self.assertTrue(build_dir.exists(), "Build directory should be created")
-
-        build_dir_files = os.listdir(str(build_dir))
-        self.assertIn("template.yaml", build_dir_files)
-        self.assertIn(function_logical_id, build_dir_files)
-
-        template_path = build_dir.joinpath("template.yaml")
-        resource_artifact_dir = build_dir.joinpath(function_logical_id)
-
-        # Make sure the template has correct CodeUri for resource
-        self._verify_resource_property(str(template_path), function_logical_id, "CodeUri", function_logical_id)
-
-        all_artifacts = set(os.listdir(str(resource_artifact_dir)))
-        actual_files = all_artifacts.intersection(expected_files)
-        self.assertEqual(actual_files, expected_files)
-
-        ruby_version = None
-        ruby_bundled_path = None
-
-        # Walk through ruby version to get to the gem path
-        for dirpath, dirname, _ in os.walk(str(resource_artifact_dir.joinpath("vendor", "bundle", "ruby"))):
-            ruby_version = dirname
-            ruby_bundled_path = Path(dirpath)
-            break
-
-        # If Gemfile is in root folder, vendor folder will also be created there
-        if ruby_version is None:
-            for dirpath, dirname, _ in os.walk(str(Path(self.working_dir).joinpath("vendor", "bundle", "ruby"))):
-                ruby_version = dirname
-                ruby_bundled_path = Path(dirpath)
-                break
-
-        gem_path = ruby_bundled_path.joinpath(ruby_version[0], "gems")
-
-        self.assertTrue(any([True if self.EXPECTED_RUBY_GEM in gem else False for gem in os.listdir(str(gem_path))]))
-
-
-class BuildIntegEsbuildBase(BuildIntegBase):
-    FUNCTION_LOGICAL_ID = "Function"
-    # Everything should be minifed to one line and a second line for the sourcemap mapping
-    MAX_MINIFIED_LINE_COUNT = 2
-
-    def _test_with_default_package_json(
-        self, runtime, use_container, code_uri, expected_files, handler, architecture=None, build_in_source=None
-    ):
-        overrides = self.get_override(runtime, code_uri, architecture, handler)
-        cmdlist = self.get_command_list(
-            use_container=use_container, parameter_overrides=overrides, build_in_source=build_in_source
-        )
-
-        LOG.info("Running Command: {}".format(cmdlist))
-        run_command(cmdlist, cwd=self.working_dir)
-
-        self._verify_built_artifact(
-            self.default_build_dir,
-            self.FUNCTION_LOGICAL_ID,
-            expected_files,
-        )
-
-        expected = {"body": '{"message":"hello world!"}', "statusCode": 200}
-        if not SKIP_DOCKER_TESTS and architecture == X86_64:
-            # ARM64 is not supported yet for invoking
-            self._verify_invoke_built_function(
-                self.built_template, self.FUNCTION_LOGICAL_ID, self._make_parameter_override_arg(overrides), expected
-            )
-
-        self._verify_esbuild_properties(self.default_build_dir, self.FUNCTION_LOGICAL_ID, handler)
-
-        if use_container:
-            self.verify_docker_container_cleanedup(runtime)
-            self.verify_pulled_image(runtime, architecture)
-
-    def _test_with_various_properties(self, overrides):
-        overrides = self.get_override(**overrides)
-        cmdlist = self.get_command_list(parameter_overrides=overrides)
-
-        LOG.info("Running Command: {}".format(cmdlist))
-        run_command(cmdlist, cwd=self.working_dir)
-
-        expected = {"body": '{"message":"hello world!"}', "statusCode": 200}
-        if not SKIP_DOCKER_TESTS and overrides["Architectures"] == X86_64:
-            # ARM64 is not supported yet for invoking
-            self._verify_invoke_built_function(
-                self.built_template, self.FUNCTION_LOGICAL_ID, self._make_parameter_override_arg(overrides), expected
-            )
-
-        self._verify_esbuild_properties(self.default_build_dir, self.FUNCTION_LOGICAL_ID, overrides["Handler"])
-
-    def _verify_esbuild_properties(self, build_dir, function_logical_id, handler):
-        filename = self._extract_filename_from_handler(handler)
-        resource_artifact_dir = build_dir.joinpath(function_logical_id)
-        self._verify_sourcemap_created(filename, resource_artifact_dir)
-        self._verify_function_minified(filename, resource_artifact_dir)
-
-    @staticmethod
-    def _extract_filename_from_handler(handler):
-        # Takes a handler in the form /a/b/c/file.function and returns file
-        return str(Path(handler).stem)
-
-    def _verify_function_minified(self, filename, resource_artifact_dir):
-        with open(Path(resource_artifact_dir, f"{filename}.js"), "r") as handler_file:
-            x = len(handler_file.readlines())
-        self.assertLessEqual(x, self.MAX_MINIFIED_LINE_COUNT)
-
-    def _verify_sourcemap_created(self, filename, resource_artifact_dir):
-        all_artifacts = set(os.listdir(str(resource_artifact_dir)))
-        self.assertIn(f"{filename}.js.map", all_artifacts)
-
-    def _verify_built_artifact(self, build_dir, function_logical_id, expected_files):
-        self.assertTrue(build_dir.exists(), "Build directory should be created")
-
-        build_dir_files = os.listdir(str(build_dir))
-        self.assertIn("template.yaml", build_dir_files)
-        self.assertIn(function_logical_id, build_dir_files)
-
-        template_path = build_dir.joinpath("template.yaml")
-        resource_artifact_dir = build_dir.joinpath(function_logical_id)
-
-        # Make sure the template has correct CodeUri for resource
-        self._verify_resource_property(str(template_path), function_logical_id, "CodeUri", function_logical_id)
-
-        all_artifacts = set(os.listdir(str(resource_artifact_dir)))
-        actual_files = all_artifacts.intersection(expected_files)
-        self.assertEqual(actual_files, expected_files)
-
-
-class BuildIntegNodeBase(BuildIntegBase):
-    EXPECTED_FILES_PROJECT_MANIFEST = {"node_modules", "main.js"}
-    EXPECTED_NODE_MODULES = {"minimal-request-promise"}
-
-    FUNCTION_LOGICAL_ID = "Function"
-
-    def _test_with_default_package_json(self, runtime, use_container, relative_path, architecture=None):
-        overrides = self.get_override(runtime, "Node", architecture, "ignored")
-        cmdlist = self.get_command_list(use_container=use_container, parameter_overrides=overrides)
-
-        LOG.info("Running Command: {}".format(cmdlist))
-        run_command(cmdlist, cwd=self.working_dir)
-
-        self._verify_built_artifact(
-            self.default_build_dir,
-            self.FUNCTION_LOGICAL_ID,
-            self.EXPECTED_FILES_PROJECT_MANIFEST,
-            self.EXPECTED_NODE_MODULES,
-        )
-
-        self._verify_resource_property(
-            str(self.built_template),
-            "OtherRelativePathResource",
-            "BodyS3Location",
-            os.path.relpath(
-                os.path.normpath(os.path.join(str(str(relative_path)), "SomeRelativePath")),
-                str(self.default_build_dir),
-            ),
-        )
-
-        self._verify_resource_property(
-            str(self.built_template),
-            "GlueResource",
-            "Command.ScriptLocation",
-            os.path.relpath(
-                os.path.normpath(os.path.join(str(str(relative_path)), "SomeRelativePath")),
-                str(self.default_build_dir),
-            ),
-        )
-
-        if use_container:
-            self.verify_docker_container_cleanedup(runtime)
-            self.verify_pulled_image(runtime, architecture)
-
-    def _verify_built_artifact(self, build_dir, function_logical_id, expected_files, expected_modules):
-        self.assertTrue(build_dir.exists(), "Build directory should be created")
-
-        build_dir_files = os.listdir(str(build_dir))
-        self.assertIn("template.yaml", build_dir_files)
-        self.assertIn(function_logical_id, build_dir_files)
-
-        template_path = build_dir.joinpath("template.yaml")
-        resource_artifact_dir = build_dir.joinpath(function_logical_id)
-
-        # Make sure the template has correct CodeUri for resource
-        self._verify_resource_property(str(template_path), function_logical_id, "CodeUri", function_logical_id)
-
-        all_artifacts = set(os.listdir(str(resource_artifact_dir)))
-        actual_files = all_artifacts.intersection(expected_files)
-        self.assertEqual(actual_files, expected_files)
-
-        all_modules = set(os.listdir(str(resource_artifact_dir.joinpath("node_modules"))))
-        actual_files = all_modules.intersection(expected_modules)
-        self.assertEqual(actual_files, expected_modules)
-
-
-class BuildIntegGoBase(BuildIntegBase):
-    FUNCTION_LOGICAL_ID = "Function"
-    EXPECTED_FILES_PROJECT_MANIFEST = {"hello-world"}
-
-    def _test_with_go(self, runtime, code_uri, mode, relative_path, architecture=None, use_container=False):
-        overrides = self.get_override(runtime, code_uri, architecture, "hello-world")
-        cmdlist = self.get_command_list(use_container=use_container, parameter_overrides=overrides)
-
-        # Need to pass GOPATH ENV variable to match the test directory when running build
-
-        LOG.info("Running Command: {}".format(cmdlist))
-        LOG.info("Running with SAM_BUILD_MODE={}".format(mode))
-
-        newenv = os.environ.copy()
-        if mode:
-            newenv["SAM_BUILD_MODE"] = mode
-
-        newenv["GOPROXY"] = "direct"
-        newenv["GOPATH"] = str(self.working_dir)
-
-        run_command(cmdlist, cwd=self.working_dir, env=newenv)
-
-        self._verify_built_artifact(
-            self.default_build_dir, self.FUNCTION_LOGICAL_ID, self.EXPECTED_FILES_PROJECT_MANIFEST
-        )
-
-        self._verify_resource_property(
-            str(self.built_template),
-            "OtherRelativePathResource",
-            "BodyS3Location",
-            os.path.relpath(
-                os.path.normpath(os.path.join(str(relative_path), "SomeRelativePath")),
-                str(self.default_build_dir),
-            ),
-        )
-
-        expected = "{'message': 'Hello World'}"
-        if not SKIP_DOCKER_TESTS and architecture == X86_64:
-            # ARM64 is not supported yet for invoking
-            self._verify_invoke_built_function(
-                self.built_template, self.FUNCTION_LOGICAL_ID, self._make_parameter_override_arg(overrides), expected
-            )
-
-        if use_container:
-            self.verify_docker_container_cleanedup(runtime)
-            self.verify_pulled_image(runtime, architecture)
-
-    def _verify_built_artifact(self, build_dir, function_logical_id, expected_files):
-        self.assertTrue(build_dir.exists(), "Build directory should be created")
-
-        build_dir_files = os.listdir(str(build_dir))
-        self.assertIn("template.yaml", build_dir_files)
-        self.assertIn(function_logical_id, build_dir_files)
-
-        template_path = build_dir.joinpath("template.yaml")
-        resource_artifact_dir = build_dir.joinpath(function_logical_id)
-
-        # Make sure the template has correct CodeUri for resource
-        self._verify_resource_property(str(template_path), function_logical_id, "CodeUri", function_logical_id)
-
-        all_artifacts = set(os.listdir(str(resource_artifact_dir)))
-        actual_files = all_artifacts.intersection(expected_files)
-        self.assertEqual(actual_files, expected_files)
-
-
-class BuildIntegJavaBase(BuildIntegBase):
-    FUNCTION_LOGICAL_ID = "Function"
-
-    def _test_with_building_java(
-        self,
-        runtime,
-        code_path,
-        expected_files,
-        expected_dependencies,
-        use_container,
-        relative_path,
-        architecture=None,
-    ):
-        if use_container and (SKIP_DOCKER_TESTS or SKIP_DOCKER_BUILD):
-            self.skipTest(SKIP_DOCKER_MESSAGE)
-
-        overrides = self.get_override(runtime, code_path, architecture, "aws.example.Hello::myHandler")
-        cmdlist = self.get_command_list(use_container=use_container, parameter_overrides=overrides)
-        cmdlist += ["--skip-pull-image"]
-        if code_path == self.USING_GRADLEW_PATH and use_container and IS_WINDOWS:
-            osutils.convert_to_unix_line_ending(os.path.join(self.test_data_path, self.USING_GRADLEW_PATH, "gradlew"))
-
-        LOG.info("Running Command: {}".format(cmdlist))
-        run_command(cmdlist, cwd=self.working_dir, timeout=900)
-
-        self._verify_built_artifact(
-            self.default_build_dir, self.FUNCTION_LOGICAL_ID, expected_files, expected_dependencies
-        )
-
-        self._verify_resource_property(
-            str(self.built_template),
-            "OtherRelativePathResource",
-            "BodyS3Location",
-            os.path.relpath(
-                os.path.normpath(os.path.join(str(relative_path), "SomeRelativePath")),
-                str(self.default_build_dir),
-            ),
-        )
-
-        self._verify_resource_property(
-            str(self.built_template),
-            "GlueResource",
-            "Command.ScriptLocation",
-            os.path.relpath(
-                os.path.normpath(os.path.join(str(relative_path), "SomeRelativePath")),
-                str(self.default_build_dir),
-            ),
-        )
-
-        # If we are testing in the container, invoke the function as well. Otherwise we cannot guarantee docker is on appveyor
-        if use_container:
-            expected = "Hello World"
-            if not SKIP_DOCKER_TESTS:
-                self._verify_invoke_built_function(
-                    self.built_template,
-                    self.FUNCTION_LOGICAL_ID,
-                    self._make_parameter_override_arg(overrides),
-                    expected,
-                )
-
-            self.verify_docker_container_cleanedup(runtime)
-            self.verify_pulled_image(runtime, architecture)
-
-    def _verify_built_artifact(self, build_dir, function_logical_id, expected_files, expected_modules):
-        self.assertTrue(build_dir.exists(), "Build directory should be created")
-
-        build_dir_files = os.listdir(str(build_dir))
-        self.assertIn("template.yaml", build_dir_files)
-        self.assertIn(function_logical_id, build_dir_files)
-
-        template_path = build_dir.joinpath("template.yaml")
-        resource_artifact_dir = build_dir.joinpath(function_logical_id)
-
-        # Make sure the template has correct CodeUri for resource
-        self._verify_resource_property(str(template_path), function_logical_id, "CodeUri", function_logical_id)
-
-        all_artifacts = set(os.listdir(str(resource_artifact_dir)))
-        actual_files = all_artifacts.intersection(expected_files)
-        self.assertEqual(actual_files, expected_files)
-
-        lib_dir_contents = set(os.listdir(str(resource_artifact_dir.joinpath("lib"))))
-        self.assertEqual(lib_dir_contents, expected_modules)
-
-
-class BuildIntegPythonBase(BuildIntegBase):
-    EXPECTED_FILES_PROJECT_MANIFEST = {
-        "__init__.py",
-        "main.py",
-        "numpy",
-        # 'cryptography',
-        "requirements.txt",
-    }
-
-    FUNCTION_LOGICAL_ID = "Function"
-    prop = "CodeUri"
-
-    def _test_with_default_requirements(
-        self,
-        runtime,
-        codeuri,
-        use_container,
-        relative_path,
-        do_override=True,
-        check_function_only=False,
-        architecture=None,
-    ):
-        if use_container and (SKIP_DOCKER_TESTS or SKIP_DOCKER_BUILD):
-            self.skipTest(SKIP_DOCKER_MESSAGE)
-        overrides = self.get_override(runtime, codeuri, architecture, "main.handler") if do_override else None
-        cmdlist = self.get_command_list(use_container=use_container, parameter_overrides=overrides)
-
-        LOG.info("Running Command: {}".format(cmdlist))
-        run_command(cmdlist, cwd=self.working_dir)
-
-        self._verify_built_artifact(
-            self.default_build_dir, self.FUNCTION_LOGICAL_ID, self.EXPECTED_FILES_PROJECT_MANIFEST
-        )
-
-        if not check_function_only:
-            self._verify_resource_property(
-                str(self.built_template),
-                "OtherRelativePathResource",
-                "BodyS3Location",
-                os.path.relpath(
-                    os.path.normpath(os.path.join(str(relative_path), "SomeRelativePath")),
-                    str(self.default_build_dir),
-                ),
-            )
-
-            self._verify_resource_property(
-                str(self.built_template),
-                "GlueResource",
-                "Command.ScriptLocation",
-                os.path.relpath(
-                    os.path.normpath(os.path.join(str(relative_path), "SomeRelativePath")),
-                    str(self.default_build_dir),
-                ),
-            )
-
-            self._verify_resource_property(
-                str(self.built_template),
-                "ExampleNestedStack",
-                "TemplateURL",
-                "https://s3.amazonaws.com/examplebucket/exampletemplate.yml",
-            )
-
-        expected = {"pi": "3.14"}
-        if not SKIP_DOCKER_TESTS:
-            self._verify_invoke_built_function(
-                self.built_template,
-                self.FUNCTION_LOGICAL_ID,
-                self._make_parameter_override_arg(overrides) if do_override else None,
-                expected,
-            )
-            if use_container:
-                self.verify_docker_container_cleanedup(runtime)
-                self.verify_pulled_image(runtime, architecture)
-
-    def _verify_built_artifact(self, build_dir, function_logical_id, expected_files):
-        self.assertTrue(build_dir.exists(), "Build directory should be created")
-
-        build_dir_files = os.listdir(str(build_dir))
-        self.assertIn("template.yaml", build_dir_files)
-        self.assertIn(function_logical_id, build_dir_files)
-
-        template_path = build_dir.joinpath("template.yaml")
-        resource_artifact_dir = build_dir.joinpath(function_logical_id)
-
-        # Make sure the template has correct CodeUri for resource
-        self._verify_resource_property(str(template_path), function_logical_id, self.prop, function_logical_id)
-
-        all_artifacts = set(os.listdir(str(resource_artifact_dir)))
-        actual_files = all_artifacts.intersection(expected_files)
-        self.assertEqual(actual_files, expected_files)
-
-    def _get_python_version(self):
-        return "python{}.{}".format(sys.version_info.major, sys.version_info.minor)
-
-
-class BuildIntegProvidedBase(BuildIntegBase):
-    EXPECTED_FILES_PROJECT_MANIFEST = {"__init__.py", "main.py", "requests", "requirements.txt"}
-
-    FUNCTION_LOGICAL_ID = "Function"
-
-    def _test_with_Makefile(
-        self, runtime, use_container, manifest, architecture=None, code_uri="Provided", build_in_source=None
-    ):
-        if use_container and (SKIP_DOCKER_TESTS or SKIP_DOCKER_BUILD):
-            self.skipTest(SKIP_DOCKER_MESSAGE)
-
-        overrides = self.get_override(runtime, code_uri, architecture, "main.handler")
-        manifest_path = None
-        if manifest:
-            manifest_path = os.path.join(self.test_data_path, "Provided", manifest)
-
-        cmdlist = self.get_command_list(
-            use_container=use_container,
-            parameter_overrides=overrides,
-            manifest_path=manifest_path,
-            build_in_source=build_in_source,
-        )
-
-        LOG.info("Running Command: {}".format(cmdlist))
-        # Built using Makefile for a python project.
-        run_command(cmdlist, cwd=self.working_dir)
-
-        if self.is_nested_parent:
-            self._verify_built_artifact_in_subapp(
-                self.default_build_dir, "SubApp", self.FUNCTION_LOGICAL_ID, self.EXPECTED_FILES_PROJECT_MANIFEST
-            )
-        else:
-            self._verify_built_artifact(
-                self.default_build_dir, self.FUNCTION_LOGICAL_ID, self.EXPECTED_FILES_PROJECT_MANIFEST
-            )
-
-        expected = "2.23.0"
-        # Building was done with a makefile, but invoke should be checked with corresponding python image.
-        overrides["Runtime"] = self._get_python_version()
-        if not SKIP_DOCKER_TESTS:
-            self._verify_invoke_built_function(
-                self.built_template, self.FUNCTION_LOGICAL_ID, self._make_parameter_override_arg(overrides), expected
-            )
-        if use_container:
-            self.verify_docker_container_cleanedup(runtime)
-            self.verify_pulled_image(runtime, architecture)
-
-    def _verify_built_artifact(self, build_dir, function_logical_id, expected_files):
-        self.assertTrue(build_dir.exists(), "Build directory should be created")
-
-        build_dir_files = os.listdir(str(build_dir))
-        self.assertIn("template.yaml", build_dir_files)
-        self.assertIn(function_logical_id, build_dir_files)
-
-        template_path = build_dir.joinpath("template.yaml")
-        resource_artifact_dir = build_dir.joinpath(function_logical_id)
-
-        # Make sure the template has correct CodeUri for resource
-        self._verify_resource_property(str(template_path), function_logical_id, "CodeUri", function_logical_id)
-
-        all_artifacts = set(os.listdir(str(resource_artifact_dir)))
-        actual_files = all_artifacts.intersection(expected_files)
-        self.assertEqual(actual_files, expected_files)
-
-    def _verify_built_artifact_in_subapp(self, build_dir, subapp_path, function_logical_id, expected_files):
-        self.assertTrue(build_dir.exists(), "Build directory should be created")
-        subapp_build_dir = Path(build_dir, subapp_path)
-        self.assertTrue(subapp_build_dir.exists(), f"Build directory for sub app {subapp_path} should be created")
-
-        build_dir_files = os.listdir(str(build_dir))
-        self.assertIn("template.yaml", build_dir_files)
-
-        subapp_build_dir_files = os.listdir(str(subapp_build_dir))
-        self.assertIn("template.yaml", subapp_build_dir_files)
-        self.assertIn(function_logical_id, subapp_build_dir_files)
-
-        template_path = subapp_build_dir.joinpath("template.yaml")
-        resource_artifact_dir = subapp_build_dir.joinpath(function_logical_id)
-
-        # Make sure the template has correct CodeUri for resource
-        self._verify_resource_property(str(template_path), function_logical_id, "CodeUri", function_logical_id)
-
-        all_artifacts = set(os.listdir(str(resource_artifact_dir)))
-        actual_files = all_artifacts.intersection(expected_files)
-        self.assertEqual(actual_files, expected_files)
-
-    def _get_python_version(self):
-        return "python{}.{}".format(sys.version_info.major, sys.version_info.minor)
 
 
 class DedupBuildIntegBase(BuildIntegBase):
